finished the game over verify

+++ src/BL/CoordinatesOfForms.java
@@ -77,7 +77,7 @@ public class CoordinatesOfForms {
         switch (index)
         {
             case 0:return new Point2D[]{new Point2D.Double(0,2),new Point2D.Double(1,0),new Point2D.Double(1,1),new Point2D.Double(1,2)};
-            case 1:return new Point2D[]{new Point2D.Double(0,1),new Point2D.Double(1,0),new Point2D.Double(2,0),new Point2D.Double(2,1)};
+            case 1:return new Point2D[]{new Point2D.Double(0,0),new Point2D.Double(1,0),new Point2D.Double(2,0),new Point2D.Double(2,1)};
             case 2:return new Point2D[]{new Point2D.Double(0,0),new Point2D.Double(0,1),new Point2D.Double(0,2),new Point2D.Double(1,0)};
             case 3:return new Point2D[]{new Point2D.Double(0,0),new Point2D.Double(0,1),new Point2D.Double(1,1),new Point2D.Double(2,1)};
             default:return new Point2D[]{new Point2D.Double(0,2),new Point2D.Double(1,0),new Point2D.Double(1,1),new Point2D.Double(1,2)};
